Refactor `trimEmptyParts` to not use `array_filter`
We need to use the `ARRAY_FILTER_USE_BOTH` flag if we use `array_filter`, since we need to check the index in the filtering logic; however, this flag
wasn't added until PHP 5.6, and we're supposed to be supporting back to PHP 5.4.

Instead use `foreach`, and build a filtered array by appending elements from the parts array which we want to keep.