Add PHP version constraints to Composer config
Existing support was for PHP >=5.4.
This is based on Travis config file, and commit #d52ecca which makes changes to maintain support for PHP 5.4.
Need to also add support for all current PHP 7.x as this is implicit by not having any current constraints.
Add PHP 7.x versions to Travis, but allow for failures as these were never intended to be supported,
and the current version of PHPUnit won't work on PHP >=7.1.
Also allow failures for PHP 5.4 and 5.5 as Travis no longer supports these.